refactor: move auth actions and routes to lib/features/auth
